+++ packages/configs/eslint-plugin-path-alias/rules/import-from-tsconfig-paths.js
@@ -135,9 +135,21 @@ export const importFromTsconfigPaths = createRule({
       return {}
     }
 
-    const aliases = projectTsconfigs.flatMap(({path, tsconfig}) =>
+    const projectAliases = projectTsconfigs.flatMap(({path, tsconfig}) =>
       parsePathAliases(tsconfig, dirname(path)),
     )
+    const compilerOptions =
+      context.sourceCode.parserServices.program?.getCompilerOptions()
+    const currentProjectAliases = compilerOptions?.configFilePath
+      ? parsePathAliases(
+          {compilerOptions},
+          compilerOptions.pathsBasePath ||
+            dirname(compilerOptions.configFilePath),
+        )
+      : projectAliases
+    const aliases = compilerOptions?.configFilePath
+      ? [...currentProjectAliases, ...projectAliases]
+      : projectAliases
     const currentFilePath = toPosixPath(resolvePath(fileName))
 
     function checkImport(importPath, node) {
@@ -147,7 +159,10 @@ export const importFromTsconfigPaths = createRule({
       const absoluteTarget = toPosixPath(
         resolvePath(dirname(fileName), importPath),
       )
-      const currentAliasRoot = findAliasRoot(currentFilePath, aliases)
+      const currentAliasRoot = findAliasRoot(
+        currentFilePath,
+        currentProjectAliases,
+      )
       if (!currentAliasRoot) {
         return
       }