+++ packages/docs/angular-table-docs/qui-docs.config.ts
@@ -107,7 +107,7 @@ export default {
    */
   hotUpdateIgnore: /angular-demo-module|public/,
   knowledge: {
-    baseUrl: "https://angular-table-next.qui.qualcomm.com",
+    baseUrl: "https://angular-table.qui.qualcomm.com",
     exclude: ["installation.mdx", "index/_index.mdx"],
     integrations: {
       openWebUi: [{id: "qui-ai"}, {id: "saga-ai"}],