+++ packages/docs/react-table-docs/qui-docs.config.ts
@@ -100,7 +100,7 @@ const navConfig: NavConfig[] = [
 export default {
   appDirectory: "src",
   knowledge: {
-    baseUrl: "https://react-table-next.qui.qualcomm.com",
+    baseUrl: "https://react-table.qui.qualcomm.com",
     exclude: ["installation.mdx", "index/_index.mdx"],
     integrations: {
       openWebUi: [{id: "qui-ai"}, {id: "saga-ai"}, {id: "saga-ai-test"}],