+++ packages/frameworks/react/vite.config.ts
@@ -1,5 +1,4 @@
-import babel from "@rolldown/plugin-babel"
-import react, {reactCompilerPreset} from "@vitejs/plugin-react"
+import react from "@vitejs/plugin-react"
 import {defineConfig} from "vite"
 
 import {dependenciesToExternal, libraryEntriesPlugin} from "@qualcomm-ui/vite"
@@ -29,11 +28,5 @@ export default defineConfig({
     },
     sourcemap: true,
   },
-  plugins: [
-    react(),
-    babel({
-      presets: [reactCompilerPreset()],
-    }),
-    libraryEntriesPlugin(),
-  ],
+  plugins: [react({compiler: true}), libraryEntriesPlugin()],
 })